+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ### 28 January 2020
 
 * Added fastlane steps for exporting & importing translations.
+* Some fixes for bootstrap script (path issues).
 
 ### 21 January 2020